fix: support plain text
The piece of code that was supposed to generate a plain text fragment
was calling buildTextElement (which doesn't work) but not even with
the right arguments: (xmlFragment, text) instead of (text)!

This feature is obviously not maintained but we need it.

The escape-html dependency is now unused, which is very
surprising. The commit introducing it didn't give any explanation.